% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_ENet.R
\name{pared_ENet}
\alias{pared_ENet}
\title{GP-Based Pareto Front for Elastic-Net}
\usage{
pared_ENet(
  X,
  y,
  lb = c(0, -6),
  ub = c(1, 6),
  Pareto_budget = 100,
  plot_title = "Elastic-Net",
  plot_marker_color = "rgba(255, 0, 0, 0.7)",
  plot_marker_border_color = "black",
  plot_marker_size = 10,
  plot_marker_border_width = 2,
  plot_marker_symbol = "circle",
  fontsize_x = 20,
  fontsize_y = 20,
  fontsize_z = 20,
  plot_title_size = 40,
  title_pos_x = 0.5,
  title_pos_y = 0.95,
  draw_projection = 1,
  proj_line_color = "black",
  proj_line_width = 2,
  proj_marker_size = 2,
  proj_marker_color = "rgba(255, 0, 0, 0.7)",
  proj_marker_border_width = 2,
  proj_marker_border_color = "black"
)
}
\arguments{
\item{X}{A numeric matrix of predictor variables (n x p), where n is the number
of samples and p is the number of predictors.}

\item{y}{A numeric vector of length n representing the response variable.}

\item{lb}{A numeric vector specifying the lower bounds for the parameters: α (\code{alpha}) and log₁₀(λ) (\code{lambda}). Default is c(0, -6).}

\item{ub}{A numeric vector specifying the upper bounds for the parameters: α (\code{alpha}) and log₁₀(λ) (\code{lambda}). Default is c(1, 6).}

\item{Pareto_budget}{Number of evaluations in the GP-based Pareto optimization. Default is 100.}

\item{plot_title}{A character string for the title of the plot. Default is "Elastic-Net".}

\item{plot_marker_color}{Fill color of markers. Default is red with transparency.}

\item{plot_marker_border_color}{Border color of markers. Default is black.}

\item{plot_marker_size}{Size of Pareto front markers. Default value is 10.}

\item{plot_marker_border_width}{Width of marker borders. Default value is 2.}

\item{plot_marker_symbol}{Symbol shape for plot markers (e.g., 'circle', 'square', 'diamond'). Default shape is 'circle'.}

\item{fontsize_x}{Font size of the x-axis title. Default value is 20.}

\item{fontsize_y}{Font size of the y-axis title. Default value is 20.}

\item{fontsize_z}{Font size of the z-axis title. Default value is 20.}

\item{plot_title_size}{Font size for the plot title. Default value is 40.}

\item{title_pos_x}{Horizontal position of the title (0 to 1 scale). Default is 0.5.}

\item{title_pos_y}{Vertical position of the title (0 to 1 scale). Default is 0.95.}

\item{draw_projection}{Logical flag (0 or 1) to add projections from each point to the XY plane. Default is 1.}

\item{proj_line_color}{Color for projection lines. Default is black.}

\item{proj_line_width}{Width of projection lines. Default value is 2.}

\item{proj_marker_size}{Size of projection plane markers. Default value is 2.}

\item{proj_marker_color}{Fill color of projection markers. Default is red with transparency.}

\item{proj_marker_border_width}{Border width of projection markers. Default value is 2.}

\item{proj_marker_border_color}{Border color of projection markers. Default is black.}
}
\value{
A list with two components:
\itemize{
\item \code{figure}: A \code{plotly} 3D scatter plot object showing the Pareto front.
\item \code{summary_table}: A data frame containing the summary of the Pareto
optimal solutions, including the values for \eqn{\alpha}, \eqn{\log10} of \eqn{\lambda},
number of non-zero coefficients, L2 norm, and deviance.
}
}
\description{
This function performs a Pareto optimality search on the Elastic-Net problem
using GP-based optimization (\code{easyGParetoptim}) and creates a 3D scatter plot
of the solution space. The plot displays the number of non-zero coefficients,
L2 norm of coefficients, and deviance for different Pareto-optimal combinations
of the regularization parameters \eqn{\alpha} (0 < \eqn{\alpha} < 1) and \eqn{\lambda}.
}
\details{
The Elastic Net objective function optimized in this process is:
\deqn{
\mathcal{L}(\beta) = \frac{1}{2n} \|y - X\beta\|_2^2 + \lambda \left[ \alpha \|\beta\|_1 + (1 - \alpha) \|\beta\|_2^2 \right]
}
where \eqn{X} is the design matrix, \eqn{y} is the response vector,
\eqn{\beta} is the coefficient vector, \eqn{\lambda > 0} is the regularization strength,
and \eqn{\alpha \in [0, 1]} balances Lasso and Ridge penalties.

This function performs Gaussian Process (GP)-based Pareto optimization to find and visualize optimal Elastic Net models with respect to multiple objectives: sparsity (number of non-zero coefficients), L2 norm of coefficients, and deviance.

The function uses \code{glmnet} for fitting Elastic Net models and \code{easyGParetoptim} for Pareto optimization.
Each point in the plot corresponds to a model, where:
\itemize{
\item x-axis: Number of non-zero coefficients (sparsity).
\item y-axis: L2 norm of the coefficients (shrinkage).
\item z-axis: Deviance (model fit).
}
The lower these three quantities, the better the model according to their respective criteria.
}
\examples{
\dontrun{
set.seed(1)
p <- 5
X <- matrix(rnorm(100), ncol = p)
n <- dim(X)[1]
beta.true <- matrix(c(1,2), ncol = 1)  # only first few coordinates are non-zero
y <- X[, 1:2] \%*\% beta.true + rnorm(n)

A <- pared_ENet(X, y, Pareto_budget = 50)
A$fig
A$summary_table
}

}
