% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_JGL.R
\name{pared_JGL}
\alias{pared_JGL}
\title{GP-Based Pareto Front for Joint Graphical Lasso}
\usage{
pared_JGL(
  sample_list,
  lb = c(-2, -2),
  ub = c(0, 0),
  Pareto_budget = 100,
  method = "group",
  plot_title = NULL,
  plot_marker_color = "rgba(255, 0, 0, 0.7)",
  plot_marker_border_color = "black",
  plot_marker_size = 10,
  plot_marker_border_width = 2,
  plot_marker_symbol = "circle",
  fontsize_x = 20,
  fontsize_y = 20,
  fontsize_z = 20,
  plot_title_size = 40,
  title_pos_x = 0.5,
  title_pos_y = 0.95,
  draw_projection = 1,
  proj_line_color = "black",
  proj_line_width = 2,
  proj_marker_size = 2,
  proj_marker_color = "rgba(255, 0, 0, 0.7)",
  proj_marker_border_width = 2,
  proj_marker_border_color = "black"
)
}
\arguments{
\item{sample_list}{A list of class-specific data samples used within the JGL wrapper functions.}

\item{lb}{A numeric vector specifying the lower bounds (log10-scale) for \eqn{\lambda_1} and \eqn{\lambda_2}. Default is c(-2, -2).}

\item{ub}{A numeric vector specifying the upper bounds (log10-scale) for \eqn{\lambda_1} and \eqn{\lambda_2}. Default is c(0, 0).}

\item{Pareto_budget}{Number of evaluations in the GP-based Pareto optimization. Default is 100.}

\item{method}{Type of JGL penalty. Either \code{"group"} or \code{"fused"}. Default is \code{"group"}.}

\item{plot_title}{Optional custom title for the plot. Automatically assigned if NULL.}

\item{plot_marker_color}{Fill color of markers. Default is red with transparency.}

\item{plot_marker_border_color}{Border color of markers. Default is black.}

\item{plot_marker_size}{Size of Pareto front markers. Default value is 10.}

\item{plot_marker_border_width}{Width of marker borders. Default value is 2.}

\item{plot_marker_symbol}{Symbol shape for plot markers (e.g., 'circle', 'square', 'diamond'). Default shape is 'circle'.}

\item{fontsize_x}{Font size of the x-axis title. Default value is 20.}

\item{fontsize_y}{Font size of the y-axis title. Default value is 20.}

\item{fontsize_z}{Font size of the z-axis title. Default value is 20.}

\item{plot_title_size}{Font size for the plot title. Default value is 40.}

\item{title_pos_x}{Horizontal position of the title (0 to 1 scale). Default is 0.5.}

\item{title_pos_y}{Vertical position of the title (0 to 1 scale). Default is 0.95.}

\item{draw_projection}{Logical flag (0 or 1) to add projections from each point to the XY plane. Default is 1.}

\item{proj_line_color}{Color for projection lines. Default is black.}

\item{proj_line_width}{Width of projection lines. Default value is 2.}

\item{proj_marker_size}{Size of projection plane markers. Default value is 2.}

\item{proj_marker_color}{Fill color of projection markers. Default is red with transparency.}

\item{proj_marker_border_width}{Border width of projection markers. Default value is 2.}

\item{proj_marker_border_color}{Border color of projection markers. Default is black.}
}
\value{
A list with two elements:
\describe{
\item{figure}{A \code{plotly} 3D scatter plot object showing the Pareto front.}
\item{summary_table}{A data frame with optimal \code{lambda1}, \code{lambda2},
and objectives, namely, total number of edges (for both group and fused), shared number of edges (group)
or MAD of Precision matrices (fused) computed pair-wise, and AIC (for both group and fused).}
}
}
\description{
Generates an interactive 3D Pareto front plot for the Joint Graphical Lasso (JGL) using
GP-based optimization (\code{easyGParetoptim}). Supports both group and fused penalties.
Additionally, it provides the list of Pareto-optimal values of the tuning parameters \eqn{\lambda_1}
and \eqn{\lambda_2}.
}
\details{
The function fits Joint Graphical Lasso using \code{JGL} and uses \code{easyGParetoptim} for Pareto optimization.
Each point in the plot corresponds to a model, where:
\itemize{
\item x-axis: Total number of edges (for both group and fused).
\item y-axis: Shared number of edges (group)/Mean absolute deviation of Precision matrices computed pair-wise (fused).
\item z-axis: Akaike Information Criteria (AIC; for both group and fused).
}
Lower values of all quantities but Shared number of edges (for that higher value is favored),
are favored while computing the Pareto-optimal points. For details on the objective functions of the Joint
Graphical Lasso with 'group' and 'fused' penalties, see Danaher et al. (2014), linked below.
}
\examples{
\dontrun{
sample_data <- generate_sample(sample_sizes = c(30, 50, 40, 70), rand_seed = 123)
result <- pared_JGL(sample_list = sample_data, method = "group", Pareto_budget = 50)
result$figure
result$summary_table

sample_data2 <- generate_sample()
resultFused <- pared_JGL(sample_list = sample_data2, method = "fused", 
plot_marker_symbol = 'diamond', plot_marker_color = 'blue', 
plot_marker_size = 7, Pareto_budget = 40)

resultFused$figure
resultFused$summary_table
}

}
\references{
\itemize{

\item Danaher P, Wang P, Witten DM. \emph{The joint graphical lasso for inverse covariance estimation across multiple classes.}\cr
J R Stat Soc Series B Stat Methodol. 2014 Mar;76(2):373-397. Doi: 10.1111/rssb.12033. \cr
(available at \url{https://pubmed.ncbi.nlm.nih.gov/24817823/}).
}
}
